requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.3
opencv-python-headless==4.5.3.56
numpy==1.21.2
gunicorn==20.1.0
